+++ wrapper/src/main/java/software/amazon/jdbc/dialect/MysqlDialect.java
@@ -24,6 +24,7 @@ import java.util.Arrays;
 import java.util.List;
 import software.amazon.jdbc.exceptions.ExceptionHandler;
 import software.amazon.jdbc.exceptions.MySQLExceptionHandler;
+import software.amazon.jdbc.hostlistprovider.ConnectionStringHostListProvider;
 
 public class MysqlDialect implements Dialect {
 
@@ -76,5 +77,9 @@ public class MysqlDialect implements Dialect {
   @Override
   public List<String> getDialectUpdateCandidates() {
     return dialectUpdateCandidates;
+  }
+
+  public HostListProviderSupplier getHostListProvider() {
+    return ConnectionStringHostListProvider::new;
   }
 }