+++ wrapper/src/main/java/software/amazon/jdbc/plugin/failover/FailoverConnectionPlugin.java
@@ -207,7 +207,7 @@ public class FailoverConnectionPlugin extends AbstractConnectionPlugin {
   public void initHostProvider(
       final String driverProtocol,
       final String initialUrl,
-      final Properties props,
+      final Properties properties,
       final HostListProviderService hostListProviderService,
       final JdbcCallable<Void, SQLException> initHostProviderFunc)
       throws SQLException {
@@ -215,7 +215,10 @@ public class FailoverConnectionPlugin extends AbstractConnectionPlugin {
         initialUrl,
         hostListProviderService,
         initHostProviderFunc,
-        () -> new AuroraHostListProvider(driverProtocol, hostListProviderService, props, initialUrl),
+        () -> this.hostListProviderService
+            .getDialect()
+            .getHostListProvider()
+            .getProvider(properties, initialUrl, hostListProviderService),
         () ->
             new ClusterAwareReaderFailoverHandler(
                 this.pluginService,
@@ -271,7 +274,7 @@ public class FailoverConnectionPlugin extends AbstractConnectionPlugin {
     LOGGER.finer(
         () -> Messages.get(
             "Failover.parameterValue",
-            new Object[] {"failoverMode", this.failoverMode}));
+            new Object[]{"failoverMode", this.failoverMode}));
   }
 
   @Override
@@ -309,7 +312,7 @@ public class FailoverConnectionPlugin extends AbstractConnectionPlugin {
       }
     }
 
-    LOGGER.fine(() -> Messages.get("Failover.invalidNode", new Object[] {currentHost}));
+    LOGGER.fine(() -> Messages.get("Failover.invalidNode", new Object[]{currentHost}));
   }
 
   private boolean isNodeStillValid(final String node, final Map<String, EnumSet<NodeChangeOptions>> changes) {
@@ -415,7 +418,7 @@ public class FailoverConnectionPlugin extends AbstractConnectionPlugin {
       LOGGER.fine(
           () -> Messages.get(
               "Failover.establishedConnection",
-              new Object[] {host}));
+              new Object[]{host}));
     } catch (final SQLException e) {
       if (this.pluginService.getCurrentConnection() != null) {
         final String msg = "Connection to "
@@ -484,8 +487,8 @@ public class FailoverConnectionPlugin extends AbstractConnectionPlugin {
   /**
    * Transfers basic session state from one connection to another.
    *
-   * @param from     The connection to transfer state from
-   * @param to       The connection to transfer state to
+   * @param from The connection to transfer state from
+   * @param to   The connection to transfer state to
    * @throws SQLException if a database access error occurs, this method is called on a closed connection, this
    *                      method is called during a distributed transaction, or this method is called during a
    *                      transaction
@@ -509,7 +512,7 @@ public class FailoverConnectionPlugin extends AbstractConnectionPlugin {
       final Class<E> exceptionClass) throws E {
     Throwable exceptionToThrow = wrapperException;
     if (originalException != null) {
-      LOGGER.finer(() -> Messages.get("Failover.detectedException", new Object[] {originalException.getMessage()}));
+      LOGGER.finer(() -> Messages.get("Failover.detectedException", new Object[]{originalException.getMessage()}));
       if (this.lastExceptionDealtWith != originalException
           && shouldExceptionTriggerConnectionSwitch(originalException)) {
         invalidateCurrentConnection();
@@ -609,13 +612,13 @@ public class FailoverConnectionPlugin extends AbstractConnectionPlugin {
 
     this.pluginService.setCurrentConnection(result.getConnection(), result.getHost());
 
-    this.pluginService.getCurrentHostSpec().removeAlias(oldAliases.toArray(new String[] {}));
+    this.pluginService.getCurrentHostSpec().removeAlias(oldAliases.toArray(new String[]{}));
     updateTopology(true);
 
     LOGGER.fine(
         () -> Messages.get(
             "Failover.establishedConnection",
-            new Object[] {this.pluginService.getCurrentHostSpec()}));
+            new Object[]{this.pluginService.getCurrentHostSpec()}));
   }
 
   protected void failoverWriter() throws SQLException {
@@ -640,7 +643,7 @@ public class FailoverConnectionPlugin extends AbstractConnectionPlugin {
     LOGGER.fine(
         () -> Messages.get(
             "Failover.establishedConnection",
-            new Object[] {this.pluginService.getCurrentHostSpec()}));
+            new Object[]{this.pluginService.getCurrentHostSpec()}));
 
     this.pluginService.refreshHostList();
   }